Module Ppx_mica

val type_generator : (Ppxlib.structure_item list, Ppxlib.module_type_declaration) Ppxlib__Deriving.Generator.t

Instantiates the PPX deriver for exprs

val type_deriver : Ppxlib.Deriving.t

Registers the PPX deriver for the expr & ty type definitions

val interp_functor_gen_wrapper : ctxt:Ppxlib.Expansion_context.Deriver.t -> Ppxlib.module_type_declaration -> Ppxlib.structure_item list

val interp_functor_gen : (Ppxlib.structure_item list, Ppxlib.module_type_declaration) Ppxlib__Deriving.Generator.t

Instantiates the PPX deriver for the Interpret functor

val interp_deriver : Ppxlib.Deriving.t

Registers the PPX deriver for the Interpret functor

val test_harness_functor_gen : (Ppxlib.structure, Ppxlib.module_type_declaration) Ppxlib__Deriving.Generator.t

val test_harness_deriver : Ppxlib.Deriving.t

val mica_module_gen : (Ppxlib.structure, Ppxlib.module_type_declaration) Ppxlib__Deriving.Generator.t

Overall generator for all the derived code: composes all the auxiliary derivers defined in this file and produces an overall module called Mica that includes the rest of the derived code

val overall_mica_deriver : Ppxlib.Deriving.t